full stamen commit with readme and images
