+++ examples/collapsible_tree/README.md
@@ -1,0 +1,14 @@
+#  Collapsible Tree
+
+
+![](collapsable-tree.png)
+
+This diagram displays a hierarchy of a series of dimensions.
+
+![](collapsable-tree.mov)
+
+**How it works**
+
+Create a look with two or more dimensions.
+
+For example, in the collapsible tree diagram featured above, you can see the nested relationship between department, category and brand in an ecommerce catalog.
