+++ examples/matrix/README.md
@@ -1,0 +1,14 @@
+#  Matrix
+
+
+![](matrix.png)
+
+This diagram creates a reorderable matrix heatmap to display correlations between dimensions on a shared measure.
+
+![](matrix.mov)
+
+**How it works**
+
+Create a look with two dimensions and one measure.
+
+For example, in the matrix featured above, you can see a heatmap of event transition counts and the various states along the rows and columns.
