+++ examples/sankey/README.md
@@ -1,0 +1,14 @@
+#  Sankey
+
+
+![](sankey.png)
+
+This diagram creates a [sankey diagram](https://en.wikipedia.org/wiki/Sankey_diagram) to display sequences of transitions.
+
+![](sankey.mov)
+
+**How it works**
+
+Create a look with any number of dimensions and one measure.
+
+For example, in the sankey diagram above, you can see event transition counts between the various sequences of states.
