+++ examples/sunburst/README.md
@@ -1,0 +1,14 @@
+#  Sunburst
+
+
+![](sunburst.png)
+
+This diagram creates a [sunburst](http://www.datavizcatalogue.com/methods/sunburst_diagram.html) to display hierarchical data in a nested structure.
+
+![](treemap.mov)
+
+**How it works**
+
+Create a look with any amount of dimensions and one measure.
+
+For example, in the sunburst featured above, you can see event counts by the hierarchical sequence of events.
