+++ examples/treemap/README.md
@@ -1,0 +1,14 @@
+#  Treemap
+
+
+![](treemap.png)
+
+This diagram creates a [treemap](https://en.wikipedia.org/wiki/Treemapping) to display hierarchical data in a nested structure.
+
+![](treemap.mov)
+
+**How it works**
+
+Create a look with any amount of dimensions and one measure.
+
+For example, in the treemap featured above, you can see event counts by the hierarchical sequence of events.
